Made fragments for the nav-bar and one for linking the js/css that is common across all views. Implemented a way to pull the characters from the database and display them on the front end

+++ src/main/java/com/tek/genshinbuildapp/model/ArtifactMainstat.java
@@ -3,10 +3,9 @@ package com.tek.genshinbuildapp.model;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -18,7 +17,7 @@ import javax.persistence.Id;
 public class ArtifactMainstat {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    long id;
+    int id;
     String statName;
     String statValue;
 }
